fix: figure 4
Attempt to fix Roman's point on not having CS-RID as part of figure. 

While CS-RID is not specifically called out by name here its base form of an Supplemental Data Service Provider (SDSP) in UTM is.

This also closes a bit of a loop hole in that UTM was not explicitly represented in the figure - which may be desired.